Create visitor.java
The Visitor pattern represents an operation to be performed on elements of an object structure. It allows you to define a new operation without changing the classes of the elements on which it operates.
More specifically, it allows objects of the Visitor class to visit the tree and traverse the target class structure via the accept method.